Move defn of dbgcheck_on out of dbgcheck.h.
That header file was not a good place for it, since we expect users to
want to turn it on and off. I like the principle that users can have all
the control they're designed to have without have to edit a
third-party's files at all -- meaning, I don't want any intended
behavior to include them editing dbgcheck.h.

The expected behavior is now that a user will either universally turn on
dbgcheck by including -Ddbgcheck as a compiler flag, or that they'll
turn it on on a per-file basis by making sure `#define dbgcheck_on` has
been effected before every `#include "dbgcheck.h"` where want the module
to be on.