Add ability to generate lock nesting graph string.
For clarity: what is generated is a string. It is a string containing
the set of directed edges of a graph. That graph captures the behavior
seen by dbgcheck in terms of which locks were in the locked state when
other locks became locked, which I'm called the nesting graph.